requirements.txt
torch>=1.8.0
torchvision
numpy
pyyaml
tqdm